+++ src/dp_matrix.h
@@ -5,7 +5,7 @@
 #include "dp_state_table.h"
 #include "dp_step.h"
 #include "eseq.h"
-#include "matrixf.h"
+#include "matrix.h"
 #include "step_matrix.h"
 #include <inttypes.h>
 
@@ -17,16 +17,16 @@ struct dp_matrix
     struct dp_state_table const* states;
     struct matrixf*              score;
     struct step_matrix*          prev_step;
-    int32_t*                     state_col;
-    uint32_t                     nstates;
+    int16_t*                     state_col;
+    uint16_t                     nstates;
 };
 
 struct dp_matrix*                   dp_matrix_create(struct dp_state_table const* states);
 void                                dp_matrix_destroy(struct dp_matrix const* matrix);
 static inline struct dp_step*       dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix,
-                                                            unsigned row, uint32_t state);
+                                                            unsigned row, uint16_t state);
 static inline struct dp_step const* dp_matrix_get_prev_step_c(struct dp_matrix const* dp_matrix,
-                                                              unsigned row, uint32_t state);
+                                                              unsigned row, uint16_t state);
 static inline float dp_matrix_get_score(struct dp_matrix const* dp_matrix, unsigned row,
                                         struct dp_step step);
 static inline void  dp_matrix_set_score(struct dp_matrix const* dp_matrix, unsigned row,
@@ -34,13 +34,13 @@ static inline void  dp_matrix_set_score(struct dp_matrix const* dp_matrix, unsig
 void                dp_matrix_setup(struct dp_matrix* matrix, struct eseq const* eseq);
 
 static inline struct dp_step* dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix,
-                                                      unsigned row, uint32_t state)
+                                                      unsigned row, uint16_t state)
 {
     return step_matrix_get_ptr(dp_matrix->prev_step, row, state);
 }
 
 static inline struct dp_step const* dp_matrix_get_prev_step_c(struct dp_matrix const* dp_matrix,
-                                                              unsigned row, uint32_t state)
+                                                              unsigned row, uint16_t state)
 {
     return step_matrix_get_ptr_c(dp_matrix->prev_step, row, state);
 }
