+++ src/dp_trans_table.h
@@ -10,48 +10,48 @@ struct state_idx;
 
 struct dp_trans_table
 {
-    uint32_t  ntrans;       /**< Number of transitions. */
+    uint16_t  ntrans;       /**< Number of transitions. */
     float*    score;        /**< Transition score. */
-    uint32_t* source_state; /**< Source state. */
-    uint32_t* offset;       /**< Maps (target state, local trans) to score
+    uint16_t* source_state; /**< Source state. */
+    uint16_t* offset;       /**< Maps (target state, local trans) to score
                               and source_state indices. */
 };
 
-int dp_trans_table_change(struct dp_trans_table* trans_tbl, uint32_t src_state, uint32_t tgt_state,
+int dp_trans_table_change(struct dp_trans_table* trans_tbl, uint16_t src_state, uint16_t tgt_state,
                           float lprob);
-struct dp_trans_table* dp_trans_table_create(struct mstate const* const* mstates, uint32_t nstates,
+struct dp_trans_table* dp_trans_table_create(struct mstate const* const* mstates, uint16_t nstates,
                                              struct state_idx* state_idx);
 void                   dp_trans_table_destroy(struct dp_trans_table const* transition);
 void                   dp_trans_table_dump(struct dp_trans_table const* trans_tbl);
-static inline uint32_t dp_trans_table_ntrans(struct dp_trans_table const* trans_tbl,
-                                             uint32_t                     tgt_state);
+static inline uint16_t dp_trans_table_ntrans(struct dp_trans_table const* trans_tbl,
+                                             uint16_t                     tgt_state);
 struct dp_trans_table* dp_trans_table_read(FILE* stream);
-static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint32_t tgt_state,
-                                         uint32_t trans);
-static inline uint32_t dp_trans_table_source_state(struct dp_trans_table const* trans_tbl,
-                                                   uint32_t tgt_state, uint32_t trans);
-static inline uint32_t dp_trans_table_total_ntrans(struct dp_trans_table const* trans_tbl);
-int dp_trans_table_write(struct dp_trans_table const* trans, uint32_t nstates, FILE* stream);
+static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint16_t tgt_state,
+                                         uint16_t trans);
+static inline uint16_t dp_trans_table_source_state(struct dp_trans_table const* trans_tbl,
+                                                   uint16_t tgt_state, uint16_t trans);
+static inline uint16_t dp_trans_table_total_ntrans(struct dp_trans_table const* trans_tbl);
+int dp_trans_table_write(struct dp_trans_table const* trans, uint16_t nstates, FILE* stream);
 
-static inline uint32_t dp_trans_table_ntrans(struct dp_trans_table const* trans_tbl,
-                                             uint32_t                     tgt_state)
+static inline uint16_t dp_trans_table_ntrans(struct dp_trans_table const* trans_tbl,
+                                             uint16_t                     tgt_state)
 {
     return trans_tbl->offset[tgt_state + 1] - trans_tbl->offset[tgt_state];
 }
 
-static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint32_t tgt_state,
-                                         uint32_t trans)
+static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint16_t tgt_state,
+                                         uint16_t trans)
 {
     return trans_tbl->score[trans_tbl->offset[tgt_state] + trans];
 }
 
-static inline uint32_t dp_trans_table_source_state(struct dp_trans_table const* trans_tbl,
-                                                   uint32_t tgt_state, uint32_t trans)
+static inline uint16_t dp_trans_table_source_state(struct dp_trans_table const* trans_tbl,
+                                                   uint16_t tgt_state, uint16_t trans)
 {
     return trans_tbl->source_state[trans_tbl->offset[tgt_state] + trans];
 }
 
-static inline uint32_t dp_trans_table_total_ntrans(struct dp_trans_table const* trans_tbl)
+static inline uint16_t dp_trans_table_total_ntrans(struct dp_trans_table const* trans_tbl)
 {
     return trans_tbl->ntrans;
 }
